Fixed J-Unit Test
